Design and Analysis of TBCC Engine
This was the end of semester project for my Jet and Rocket Propulsion class that I took in Spring 2022. 

This code analyizes the preformance of a Turbine Based Combined Cycle engine that powers an aircraft. The three distinct operating regimes are turbojet, turbojet with afterburner, and ramjet. These three are indpendent of each other and only switch when the TSFC from the next phase is lower than the previous phase. This allows the TBCC to be the most efficient it can be. The turbojet and ramjet are both Brayton cycles.

The report was also included, which had the complete analysis and findings of the project. The report also details why certain design choices were made.