Orbit Tracking for HIAD Applications (Code and Report)
This is the code for the Orbital Tracking project I completed in my Aerospace Vehicle Performance class in the Spring of 2022. This project used several important properties and formulas from orbital mechanics and culminated into being able to find an object in a tundra orbit after a given time of flight and initial location. Several key principles used in this project include sidereal time calculations, converting the initial range and range rate to ECI vectors, calculating orbital elements of the object, and solving Kepler's and Kepler's Inverse Problem. 

'OrbitTracker.m' is the script that had the values given in the project prompt already inserted. 'OrbitTracker_2.m' is a similar code that allows users to input values for range, range rate, radar site location, time of observation, and time of flight, allowing for wider applications. This code also reduces large chunks of code by replacing them with smaller functions and scripts. These supporting files can be found in the folder.