Aerothermodynamic Heating of Hypersonic Vehicle
This code can calculate the aerodynamic coefficients for a hypersonic vehicle at various angles of attack, altitudes, and speeds. For the project, it was calculated for a vehicle with given 27 body geometry points at Mach 7 and 10, altitudes of 12.2 and 18.28 kilometers, and angles of attack from 0-10 degrees. The second part of the project looked at the stagnation point heating of the nose, as well as the skin friction heating at each body point. 

This project also utilized an Equilibrium Flow Library to calculate the flow composition and properties throughout flight.